app/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
pillow
transformers
torchvision
streamlit